Added caching of speech
Note: Generated speech will be cached in /data, you must make this path
available to your docker container.

Dockerfile changes:
Added supervisor to docker image
Added new scripts, manage_tts_cache.sh and run.sh
Changed ENTRYPOINT to run.sh

app.py changes:
Updated to check for cached speech in /data.  If it exists, it will be
returned and no new speech generated.  If it does not exist (or is
empty), then the speech is generated and saved to a file before it is
returned.  Note, too, that the cache files are named as the SHA256 hash
of the actual text, the name of the TTS engine, and the voice id.  This
means that if you request the same speech with different voices or TTS
engines, a cached result will not be used.  By using a SHA256 hash, we
also prevent exposure of the text in the cache.  If testing reveals
collisions in the SHA256 hash, dropping the TTS engine and voice id
would likely reduce collisions.

manage_tts_cache.sh:
The new manage_tts_cache.sh script is nearly identical to one I wrote
for a different project using snips.  It will delete cached speech if
the specified size is exceeded.  It gets invoked through supervisor once
an hour.  The size is hard-coded as 50000000 bytes.  This could be made
configurable.

run.sh:
The new run.sh sript is the new entrypoint for the container.  It
creates a configuration file for supervisor based on the command line
arguments (which will get passed to app.py) and runs supervisor.
Additionally, the script will wait, watching for SIGTERM and SIGKILL.
If it gets either, it will nicely terminate supervisor, which stops
app.py and manage_tts_cache.sh.

Discussion:
This should probably be modified to add some sort of configuration for
how long the data is cached.

Perhaps command-line arguments to docker could be used to configure the
size for manage_tts_cache.sh, but run.sh would need to recognize the
configuration option for manage_tts_cache.sh and pass it to
manage_tts_cache.sh while passing all other options to app.py.
Alternatively, manage_tts_cache.sh could be reworked in python, and
invoked regularly from app.py, which would eliminate the need for
run.sh and supervisor.